requirements/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements/requirements.in
#
bcrypt==3.1.7             # via passlib
cffi==1.14.0              # via bcrypt
click==7.1.1              # via uvicorn
environs==7.4.0           # via -r requirements/requirements.in
fastapi==0.54.1           # via -r requirements/requirements.in
h11==0.9.0                # via uvicorn
httptools==0.1.1          # via uvicorn
marshmallow==3.5.1        # via environs
passlib[bcrypt]==1.7.2    # via -r requirements/requirements.in
pycparser==2.20           # via cffi
pydantic==1.5.1           # via fastapi
pyjwt==1.7.1              # via -r requirements/requirements.in
python-dotenv==0.13.0     # via environs
python-multipart==0.0.5   # via -r requirements/requirements.in
six==1.14.0               # via bcrypt, python-multipart
starlette==0.13.2         # via fastapi
uvicorn==0.11.3           # via -r requirements/requirements.in
uvloop==0.14.0            # via uvicorn
websockets==8.1           # via uvicorn
